feat Front-End: Add enhanced API permissions and adjust button visibility control

+++ mini-ITS.Web/Controllers/EnrollmentsDescriptionController.cs
@@ -17,12 +17,18 @@ namespace mini_ITS.Web.Controllers
     public class EnrollmentsDescriptionController : ControllerBase
     {
         private readonly IEnrollmentsDescriptionServices _enrollmentsDescriptionServices;
+        private readonly IUsersServices _usersServices;
         private readonly IMapper _mapper;
         private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public EnrollmentsDescriptionController(IEnrollmentsDescriptionServices enrollmentsDescriptionServices, IMapper mapper, IHttpContextAccessor httpContextAccessor)
+        public EnrollmentsDescriptionController(
+            IEnrollmentsDescriptionServices enrollmentsDescriptionServices,
+            IUsersServices usersServices,
+            IMapper mapper,
+            IHttpContextAccessor httpContextAccessor)
         {
             _enrollmentsDescriptionServices = enrollmentsDescriptionServices;
+            _usersServices = usersServices;
             _mapper = mapper;
             _httpContextAccessor = httpContextAccessor;
         }
@@ -83,6 +89,17 @@ namespace mini_ITS.Web.Controllers
             try
             {
                 if (enrollmentsDescriptionDto == null) return NotFound("Error: enrollmentsDescriptionDto is null");
+
+                var userIdentity = await _usersServices.GetAsync(User.Identity.Name);
+                if (userIdentity == null) return Unauthorized("Error: user not found");
+
+                var userAddDescription = await _usersServices.GetAsync(enrollmentsDescriptionDto.UserAddDescription);
+
+                if (userIdentity.Role != "Administrator" && enrollmentsDescriptionDto.UserAddDescription != userIdentity.Id)
+                {
+                    return Forbid("Error: user not authorized to delete this enrollmentDescription");
+                }
+
                 await _enrollmentsDescriptionServices.UpdateAsync(enrollmentsDescriptionDto, User.Identity.Name);
 
                 return Ok();
@@ -99,6 +116,20 @@ namespace mini_ITS.Web.Controllers
             try
             {
                 if (id == null) return BadRequest("Error: id is null");
+
+                var userIdentity = await _usersServices.GetAsync(User.Identity.Name);
+                if (userIdentity == null) return Unauthorized("Error: user not found");
+
+                var enrollmentsDescriptionDto = await _enrollmentsDescriptionServices.GetAsync((Guid)id);
+                if (enrollmentsDescriptionDto == null) return NotFound("Error: enrollmentDescription not found");
+
+                var userAddDescription = await _usersServices.GetAsync(enrollmentsDescriptionDto.UserAddDescription);
+
+                if (userIdentity.Role != "Administrator" && enrollmentsDescriptionDto.UserAddDescription != userIdentity.Id)                       
+                {
+                    return Forbid("Error: user not authorized to delete this enrollmentDescription");
+                }
+
                 await _enrollmentsDescriptionServices.DeleteAsync((Guid)id);
 
                 return Ok();
